requirements.txt
PyQt5==5.10
pytest==3.4.1
python-can==2.1.0
